+++ src/main/ipcs/server/instance/createRemoteServerInstance.ts
@@ -1,0 +1,11 @@
+import { ipcMain } from 'electron';
+import Channels from '../../channels';
+import createRemoteServerInstance from '../../../services/remote/createRemoteServerInstance';
+import { CreateRemoteServerInstanceInput } from '../../../../types/RemoteConnection.types';
+
+ipcMain.handle(
+  Channels.createRemoteServerInstance,
+  async (event, input: CreateRemoteServerInstanceInput) => {
+    return createRemoteServerInstance(input);
+  },
+);